have it use software collections ruby
