+++ lib/ood_portal_generator/version.rb
@@ -1,4 +1,4 @@
-class OodPortalGenerator
+module OodPortalGenerator
   # The current version of OodPortalGenerator
   VERSION= "0.4.0"
 end
